requirements.prod.txt
# ARK Production Dependencies - v6.0-GENESIS-COMPLETE
# Frozen on 2025-11-11
# These are the EXACT versions tested and validated for production

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0
python-multipart==0.0.18

# Database & Caching
sqlalchemy==2.0.23
aiosqlite==0.19.0
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2

# Web Scraping
beautifulsoup4==4.12.2
playwright==1.40.0

# Data Science
pandas>=2.1.4
numpy>=1.26.0
scikit-learn>=1.3.2
matplotlib>=3.8.2
plotly==5.17.0

# Mathematics
sympy==1.12

# Telegram Bot
python-telegram-bot==20.7

# Utilities
jinja2==3.1.6
python-dotenv==1.0.0

# Cryptography
PyNaCl>=1.5.0

# Federation
websockets>=12.0

# Testing
pytest==8.3.5
pytest-asyncio==1.3.0

# ARK Core Components
# - reasoning/ - Hierarchical reasoning system
# - agents/ - Kyle, Joey, Kenny, Aletheia, ID
# - intelligent-backend.cjs - Event bus (Node.js)
# - reasoning_api.py - FastAPI integration server

# Deployment Notes:
# - Run reasoning API: python reasoning_api.py (port 8101)
# - Run backend: node intelligent-backend.cjs
# - Redis recommended but optional (pub/sub features)
# - SQLite database: data/reasoning_logs.db
